+++ resources/views/livewire/key-control/home-table.blade.php
@@ -29,8 +29,12 @@
                 <!-- Actions -->
                 <td>
                     <div class="flex flex-row flex-wrap gap-2 justify-center">
-                        <x-svg.arrow-down-button wire:click="updateKeyControlReception({{ $keyControl->id }})"/>
-                        <x-svg.recycle-bin wire:click="deleteKeyControlRecord({{ $keyControl->id }})" class="w-9 h-9 stroke-red-600 dark:stroke-red-200 bg-red-300 dark:bg-red-800 bg-opacity-40"/>
+                        @can('update', $keyControl)
+                            <x-svg.arrow-down-button wire:click="updateKeyControlReception({{ $keyControl->id }})"/>
+                        @endcan
+                        @can('cancel', $keyControl)
+                            <x-svg.recycle-bin wire:click="deleteKeyControlRecord({{ $keyControl->id }})" class="w-9 h-9 stroke-red-600 dark:stroke-red-200 bg-red-300 dark:bg-red-800 bg-opacity-40"/>
+                        @endcan
                     </div>
                 </td>
             </tr>
